move choice logic around
