+++ frontend/src/components/Inputs/OpenChoiceInput/OpenChoice.jsx
@@ -3,39 +3,35 @@ import PropTypes from 'prop-types';
 
 import './OpenChoice.css';
 import '../../ComponentStyles.css';
+import {getListOfChoices} from '../../../util/util.js';
 
 export default class OpenChoice extends Component {
     constructor(props) {
         super(props);
         this.state = {
             value: "",
-            open: false
-            
+            open: false,
+            choices:[],
+            display:""
         };
 
     this.onInputChange = this.onInputChange.bind(this);
+    this.setChoice = this.setChoice.bind(this);
     }
 
     componentDidMount() {
-        this.setState({value: this.props.options[0]})
+        // this.setState({value: this.props.options[0]})
+        getListOfChoices(this.props, this.setChoice);
     }
 
     onInputChange(event) {
         this.setState({value: event.target.value})
     }
-    formatArea(val){
-        switch(val){
-            case '<':
-                return <>&lt;</>
-            case '>':
-                return <>&gt;</>
-            case '>=':
-                return <>&ge;</>
-            case '<=':
-                return <>&le;</>
-            default:
-                return <>{val}</>
-        }
+
+    setChoice(pair) {
+        this.setState(previousState => ({
+            choices: [...previousState.choices, pair]
+        }));
     }
 
     render() {
@@ -64,16 +60,16 @@ export default class OpenChoice extends Component {
                             }
                         }}/>
                         <div className={"dropdown-block option-block " + (this.state.open?'':"hide-block")}>
-                            {this.props.options.map((e)=>{
-                            if(e!=this.state.value){
+                            {this.state.choices.map((e)=>{
+                            if(e.code!=this.state.value){
                                 return (
-                                    <div key={e} className="unselected-option" onClick={()=>{
-                                        this.setState({value:e});
+                                    <div key={e.code} className="unselected-option" onClick={()=>{
+                                        this.setState({value:e.code});
                                         this.setState({open:false})
                                     }}
                                     // prevent the dropdown from stealing focus and closing
-                                    onMouseDown={(e)=>{e.preventDefault()}}>
-                                        {this.formatArea(e)}
+                                    onMouseDown={(event)=>{event.preventDefault()}}>
+                                        {e.code}
                                     </div>
                                 )
                             }
